Move OP implementation to separate library 'pyop'.
Update implementation to work with pyop v2.0.0.
