+++ setup.py
@@ -1,13 +1,13 @@
 from setuptools import setup, find_packages
 
 setup(
-        name='se-leg-op',
-        version='0.0.1',
-        packages=find_packages('src'),
-        package_dir={'': 'src'},
-        url='https://github.com/SUNET/se-leg-op',
-        license='',
-        author='',
-        author_email='',
-        description=''
+    name='se-leg-op',
+    version='0.0.1',
+    packages=find_packages('src'),
+    package_dir={'': 'src'},
+    url='https://github.com/SUNET/se-leg-op',
+    license='',
+    author='',
+    author_email='',
+    description='',
 )
