+++ src/se_leg_op/service/run.py
@@ -1,6 +1,6 @@
 import logging
 
-from se_leg_op.service.app import oidc_provider_init_app
+from .app import oidc_provider_init_app
 
 name = 'oidc_provider'
 app = oidc_provider_init_app(name)
