+++ src/se_leg_op/service/views/oidc_provider.py
@@ -4,14 +4,10 @@ from flask import current_app
 from flask import jsonify
 from flask.helpers import make_response
 from oic.oic.message import TokenErrorResponse, UserInfoErrorResponse
+from pyop.access_token import AccessToken, BearerTokenError
+from pyop.exceptions import InvalidAuthenticationRequest, InvalidAccessToken, InvalidClientAuthentication, OAuthError
 
-from ...response_sender import deliver_response_task
-from ...access_token import AccessToken, BearerTokenError
-from ...authz_state import InvalidAccessToken
-from ...authz_state import InvalidAuthorizationCode, InvalidRefreshToken, InvalidScope
-from ...client_authentication import InvalidClientAuthentication
-from ...provider import InvalidAuthenticationRequest
-from ...provider import InvalidTokenRequest
+from ..response_sender import deliver_response_task
 
 oidc_provider_views = Blueprint('oidc_provider', __name__, url_prefix='')
 
@@ -57,12 +53,6 @@ def token_endpoint():
         token_response = current_app.provider.handle_token_request(flask.request.get_data().decode('utf-8'),
                                                                    flask.request.headers, extra_userinfo)
         return jsonify(token_response.to_dict())
-    except InvalidTokenRequest as e:
-        current_app.logger.debug('received invalid token request', exc_info=True)
-        error_resp = TokenErrorResponse(error=e.oauth_error, error_description=str(e))
-        response = make_response(error_resp.to_json(), 400)
-        response.headers['Content-Type'] = 'application/json'
-        return response
     except InvalidClientAuthentication as e:
         current_app.logger.debug('invalid client authentication at token endpoint', exc_info=True)
         error_resp = TokenErrorResponse(error='invalid_client', error_description=str(e))
@@ -70,15 +60,9 @@ def token_endpoint():
         response.headers['Content-Type'] = 'application/json'
         response.headers['WWW-Authenticate'] = 'Basic'
         return response
-    except (InvalidAuthorizationCode, InvalidRefreshToken) as e:
-        current_app.logger.debug('invalid authorization grant received', exc_info=True)
-        error_resp = TokenErrorResponse(error='invalid_grant', error_description=str(e))
-        response = make_response(error_resp.to_json(), 400)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    except InvalidScope as e:
-        current_app.logger.debug('invalid scope requested at token endpoint', exc_info=True)
-        error_resp = TokenErrorResponse(error='invalid_scope', error_description=str(e))
+    except OAuthError as e:
+        current_app.logger.debug('invalid request: %s', str(e), exc_info=True)
+        error_resp = TokenErrorResponse(error=e.oauth_error, error_description=str(e))
         response = make_response(error_resp.to_json(), 400)
         response.headers['Content-Type'] = 'application/json'
         return response
@@ -90,17 +74,13 @@ def userinfo_endpoint():
         response = current_app.provider.handle_userinfo_request(flask.request.get_data().decode('utf-8'),
                                                                 flask.request.headers)
         return jsonify(response.to_dict())
-    except InvalidAccessToken as e:
+    except (BearerTokenError, InvalidAccessToken) as e:
         error_resp = UserInfoErrorResponse(error='invalid_token', error_description=str(e))
         response = make_response(error_resp.to_json(), 401)
         response.headers['WWW-Authenticate'] = AccessToken.BEARER_TOKEN_TYPE
         response.headers['Content-Type'] = 'application/json'
         return response
-    except BearerTokenError as e:
-        response = make_response('', 401)
-        response.headers['WWW-Authenticate'] = AccessToken.BEARER_TOKEN_TYPE
-        response.headers['Content-Type'] = 'application/json'
-        return response
+
 
 def extra_userinfo(user_id, client_id):
     return {'vetting_time': current_app.provider.userinfo[user_id]['vetting_time']}
