+++ tests/requirements.txt
@@ -1,2 +1,3 @@
 pytest
 responses==0.5.1
+oic
