+++ tests/service/views/test_oidc_provider.py
@@ -11,47 +11,15 @@ import pytest
 import responses
 from jwkest.jwk import RSAKey, import_rsa_key
 from oic.oic.message import AuthorizationRequest, IdToken, ClaimsRequest, Claims
+from pyop.exceptions import InvalidAuthenticationRequest
 from rq.worker import SimpleWorker
 
-from se_leg_op.provider import InvalidAuthenticationRequest
-from se_leg_op.service.app import OpStorageWrapper
-from se_leg_op.service.app import SE_LEG_PROVIDER_SETTINGS_ENVVAR
-from se_leg_op.service.app import oidc_provider_init_app
-from tests.storage.mongodb import MongoTemporaryInstance
-from tests.storage.redis import RedisTemporaryInstance
+from se_leg_op.storage import OpStorageWrapper
 
 TEST_CLIENT_ID = 'client1'
 TEST_CLIENT_SECRET = 'secret'
 TEST_REDIRECT_URI = 'https://client.example.com/redirect_uri'
 TEST_USER_ID = 'user1'
-
-
-@pytest.yield_fixture
-def mongodb():
-    tmp_db = MongoTemporaryInstance()
-    yield tmp_db
-    tmp_db.shutdown()
-
-
-@pytest.yield_fixture
-def redis():
-    tmp_redis = RedisTemporaryInstance()
-    yield tmp_redis
-    tmp_redis.shutdown()
-
-
-@pytest.fixture
-def inject_app(request, tmpdir, mongodb, redis):
-    os.chdir(str(tmpdir))
-    os.environ[SE_LEG_PROVIDER_SETTINGS_ENVVAR] = './app_config.py'
-    config = {
-        '_mongodb': mongodb,
-        'DB_URI': mongodb.get_uri(),
-        'REDIS_URI': redis.get_uri()
-    }
-    app = oidc_provider_init_app(__name__, config=config)
-    app.authn_response_queue.empty()
-    request.instance.app = app
 
 
 @pytest.fixture
@@ -68,12 +36,12 @@ def authn_request_args():
 @pytest.mark.usefixtures('inject_app')
 class TestConfiguration(object):
     def test_config(self):
-        assert self.app.provider.provider_configuration['issuer'] == 'http://localhost:5000'
+        assert self.app.provider.provider_configuration['issuer'] == 'https://localhost:5000'
         assert self.app.provider.provider_configuration['authorization_endpoint'] == \
-               'http://localhost:5000/authentication'
-        assert self.app.provider.provider_configuration['jwks_uri'] == 'http://localhost:5000/jwks'
-        assert self.app.provider.provider_configuration['token_endpoint'] == 'http://localhost:5000/token'
-        assert self.app.provider.provider_configuration['userinfo_endpoint'] == 'http://localhost:5000/userinfo'
+               'https://localhost:5000/authentication'
+        assert self.app.provider.provider_configuration['jwks_uri'] == 'https://localhost:5000/jwks'
+        assert self.app.provider.provider_configuration['token_endpoint'] == 'https://localhost:5000/token'
+        assert self.app.provider.provider_configuration['userinfo_endpoint'] == 'https://localhost:5000/userinfo'
         assert self.app.provider.provider_configuration['scopes_supported'] == ['openid']
         assert self.app.provider.provider_configuration['response_types_supported'] == ['code', 'code id_token',
                                                                                         'code token',
@@ -108,7 +76,7 @@ class TestAuthenticationEndpoint(object):
         client_db = OpStorageWrapper(db_uri, 'clients')
         client_db[TEST_CLIENT_ID] = {
             'redirect_uris': [TEST_REDIRECT_URI],
-            'response_types': [['code']],
+            'response_types': ['code'],
         }
         self.app.provider.clients = client_db
 
@@ -372,7 +340,7 @@ class TestProviderConfigurationEndpoint(object):
     def test_configuration_endpoint(self):
         resp = self.app.test_client().get('/.well-known/openid-configuration')
         assert resp.status_code == 200
-        assert json.loads(resp.data.decode('utf-8')) == self.app.provider.provider_configuration
+        assert json.loads(resp.data.decode('utf-8')) == self.app.provider.provider_configuration.to_dict()
 
 
 @pytest.mark.usefixtures('inject_app')
