+++ tests/service/views/test_vetting_process.py
@@ -1,47 +1,16 @@
-import os
 from urllib.parse import parse_qsl, urlparse
 
 import pytest
 import responses
 from rq.worker import SimpleWorker
 
-from se_leg_op.service.app import oidc_provider_init_app, SE_LEG_PROVIDER_SETTINGS_ENVVAR, OpStorageWrapper
-from tests.storage.mongodb import MongoTemporaryInstance
-from tests.storage.redis import RedisTemporaryInstance
+from se_leg_op.storage import OpStorageWrapper
 
 TEST_CLIENT_ID = 'client1'
 TEST_CLIENT_SECRET = 'secret'
 TEST_REDIRECT_URI = 'https://client.example.com/redirect_uri'
 
 TEST_USER_ID = 'user1'
-
-
-@pytest.yield_fixture
-def mongodb():
-    tmp_db = MongoTemporaryInstance()
-    yield tmp_db
-    tmp_db.shutdown()
-
-
-@pytest.yield_fixture
-def redis():
-    tmp_redis = RedisTemporaryInstance()
-    yield tmp_redis
-    tmp_redis.shutdown()
-
-
-@pytest.fixture
-def inject_app(request, tmpdir, mongodb, redis):
-    os.chdir(str(tmpdir))
-    os.environ[SE_LEG_PROVIDER_SETTINGS_ENVVAR] = './app_config.py'
-    config = {
-        '_mongodb': mongodb,
-        'DB_URI': mongodb.get_uri(),
-        'REDIS_URI': redis.get_uri()
-    }
-    app = oidc_provider_init_app(__name__, config=config)
-    app.authn_response_queue.empty()
-    request.instance.app = app
 
 
 @pytest.fixture
@@ -65,7 +34,7 @@ class TestVettingResultEndpoint(object):
         client_db[TEST_CLIENT_ID] = {
             'redirect_uris': [TEST_REDIRECT_URI],
             'client_secret': TEST_CLIENT_SECRET,
-            'response_types': [['code']]
+            'response_types': ['code']
         }
         self.app.provider.clients = client_db
